Add basic watch page
At the moment, all the watch page has is a back button and the video
player. Of course, video details will be added. Right now, this just
uses a YouTube embed, but an experimental MPV-based player is included
and can be toggled on in the .pro file. This MPV-based player is the one
I plan on using, but I'm having problems implementing it.