Disable Windows 8 dark theme code on MSVC
The dark theme will still work on Windows 8 on MSVC builds as it's
implemented completely on the Qt UI level, however the program will not
be recognized as dark by the system and will not have a dark title bar.
This was done because MSVC does not support inline assembly, and the
solution not using it is jank and I don't like it, so it's out.