Improve TubeLabel hover logic
The hover logic has been changed to only take into account the bounding
rectangle of the text iself, rather than the rectangle of the entire
label. An I-cursor will also now be shown on hover for any non-clickable
TubeLabel. The bounding rectangle stored is not changed on resize yet, I
couldn't figure it out but I'm gonna be off of this project for a bit I
think so might as well commit now.