Inject block ads setting faster
On rare occasion, pre-roll ads will still get served with the block ads
setting on. My theory to why this happens is that the setting doesn't
get injected fast enough. QWebChannel take sa little while to load in,
and that little while might be long enough for the player's ad code to
load first. I guess we'll see if my theory's correct.