Drop Qt 5 workflow and general Qt 5 support
Qt 5, and primarily its outdated web engine, has had problems with the
program for a while. It is buggy and sometimes the program will just not
work (I believe Verdigris is to blame for the latter part). Therefore, I
don't feel support should be provided for it anymore. All the code for
compatibility that is there right now will remain there to create less
work for anyone wanting to build for Qt 5 themselves, but I'm not
keeping up with it anymore.