Remove unneeded code, respect image cache setting again
Big step in getting rid of the program's deep intertwining with the
Innertube library. There is still more to be removed, as players,
channels, authentication, and URL resolution still have to be migrated
to the plugin interface. Once those are done, most if not all of the
remaining Innertube code in the program can be stripped out.